refactor: FloatingButton 컴포넌트 molecules 디렉터리로 이동
- position: fixed 스타일을 적용하기 위해 Fixed 컴포넌트 활용
