--- src/components/@shared/FloatingButton/FloatingButton.style.ts
@@ -1,11 +1,0 @@
-import styled from '@emotion/styled';
-
-const Wrapper = styled.button`
-  position: fixed;
-  right: 50px;
-  bottom: 50px;
-  background-color: transparent;
-  cursor: pointer;
-`;
-
-export { Wrapper };
