--- src/components/@shared/FloatingButton/FloatingButton.type.ts
@@ -1,9 +1,0 @@
-import { SerializedStyles } from '@emotion/react';
-import { MouseEventHandler } from 'react';
-
-type FloatingButtonProps = {
-  customStyle?: SerializedStyles;
-  onClick?: MouseEventHandler<HTMLButtonElement>;
-};
-
-export { FloatingButtonProps };
