+++ src/components/molecules/FloatingButton/FloatingButton.stories.tsx
@@ -1,21 +1,36 @@
 import { css } from '@emotion/react';
 import styled from '@emotion/styled';
-import FloatingButton from 'components/@shared/FloatingButton/FloatingButton';
+import FloatingButton from './FloatingButton';
 
 export default {
-  title: '@shared/FloatingButton',
+  title: 'molecules/FloatingButton',
   component: FloatingButton,
 };
 
 export const DefaultFloatingButton = () => (
   <FloatingButton
-    customStyle={css`
-      opacity: 0.6;
-    `}
+    layer={FLOATING_BUTTON_LAYER}
+    wrapperCustomStyle={wrapperCustomStyle}
+    buttonCustomStyle={buttonCustomStyle}
   >
     <ButtonInnerWrapper>버튼</ButtonInnerWrapper>
   </FloatingButton>
 );
+
+const wrapperCustomStyle = css`
+  right: 50px;
+  bottom: 50px;
+`;
+
+const buttonCustomStyle = css`
+  background-color: transparent;
+  transition: 0.5s;
+  opacity: 0.6;
+
+  &:hover {
+    opacity: 1;
+  }
+`;
 
 const ButtonInnerWrapper = styled.div`
   ${({ theme }) => css`
@@ -30,3 +45,5 @@ const ButtonInnerWrapper = styled.div`
     text-align: center;
   `}
 `;
+
+const FLOATING_BUTTON_LAYER = 50;
