+++ src/components/molecules/FloatingButton/FloatingButton.style.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+const Button = styled.button`
+  cursor: pointer;
+`;
+
+export { Button };
